Linux: Allow set custom baudrate
Taken from https://github.com/renathoaz/go-serial/commit/360227639015e7cb8e025c3fe9c044fe0c724e46
Implementend ONLY for Linux
