+++ constants_darwin.go
