+++ constants_darwin_386.go
