+++ constants_darwin_amd64.go
