+++ constants_linux.go
@@ -57,13 +57,6 @@ var databitsMap = map[int]uint32{
 
 const tcCMSPAR = unix.CMSPAR
 const tcIUCLC = unix.IUCLC
-
 const tcCRTSCTS uint32 = unix.CRTSCTS
 
-const ioctlTcgetattr = unix.TCGETS
-const ioctlTcsetattr = unix.TCSETS
 const ioctlTcflsh = unix.TCFLSH
-
-func toTermiosSpeedType(speed uint32) uint32 {
-	return speed
-}
