+++ termios_baudrate_unix.go
@@ -1,0 +1,29 @@
+//
+// Copyright 2014-2017 Cristian Maglie. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+//
+
+// +build darwin freebsd openbsd
+
+package serial
+
+import (
+	"golang.org/x/sys/unix"
+)
+
+func setTermSettingsBaudrate(speed int, settings *unix.Termios) error {
+	baudrate, ok := baudrateMap[speed]
+	if !ok {
+		return &PortError{code: InvalidSpeed}
+	}
+	// revert old baudrate
+	for _, rate := range baudrateMap {
+		settings.Cflag &^= rate
+	}
+	// set new baudrate
+	settings.Cflag |= baudrate
+	settings.Ispeed = toTermiosSpeedType(baudrate)
+	settings.Ospeed = toTermiosSpeedType(baudrate)
+	return nil
+}
